X hair added, mouse movement also. looking good

+++ src/RayCastCore/MovingObject.java
@@ -3,8 +3,8 @@ package RayCastCore;
 import java.util.ArrayList;
 
 public class MovingObject {
-    public float positionX;
-    public float positionY;
+    public double positionX;
+    public double positionY;
     public float direction = 0;//degrees
     public int size;
     public int health = 1;
@@ -13,8 +13,8 @@ public class MovingObject {
     public int speed;
 
     public void collisionCheck(ArrayList<LineSegment> segments, double changeX, double changeY){
-        float startX = positionX;
-        float startY = positionY;
+        double startX = positionX;
+        double startY = positionY;
         positionX += (float) changeX;
         positionY += (float) changeY;
