+++ src/RayCastCore/RayCast.java
@@ -11,7 +11,7 @@ public class RayCast {
 
     public static double distanceEff(Point a, Point b){return Math.pow(b.x-a.x,2)+Math.pow(b.y-a.y,2);}
 
-    public static double distanceEff(float x1, float y1, float x2, float y2){return Math.pow(x1-x2,2)+Math.pow(y1-y2,2);}
+    public static double distanceEff(double x1, double y1, double x2, double y2){return Math.pow(x1-x2,2)+Math.pow(y1-y2,2);}
 
     public static Point intersectLines(LineSegment change, LineSegment wall){
         float a = (float) change.angleGrad;
@@ -67,7 +67,7 @@ public class RayCast {
         return null;
     }
 
-    protected static Point intersectCircleRay(LineSegment ray, float positionX, float positionY, int size, int type) {
+    protected static Point intersectCircleRay(LineSegment ray, double positionX, double positionY, int size, int type) {
         double perpendicular = ray.angleRad - (Math.PI / 2); //-90 degrees basically to get the perpendicular
         double perpX = size / 2 * Math.cos(perpendicular);
         double perpY = size / 2 * Math.sin(perpendicular);
@@ -85,7 +85,7 @@ public class RayCast {
         return null;
     }
 
-    public static boolean CirclesCollision(float positionX1, float positionY1, int size1, float positionX2, float positionY2, int size2) {
+    public static boolean CirclesCollision(double positionX1, double positionY1, int size1, double positionX2, double positionY2, int size2) {
         return distanceEff(positionX1, positionY1, positionX2, positionY2) <= (Math.pow(size1/2 + size2/2,2));
     }
